Add Cache-Control in Umbra::RequestBuilder
Add `Cache-Control: no-cache, no-store, private, max-age=0` to umbra
requests.

We want to ensure that we bypass any cache in order to properly shadow a
request. If an initial request missed cache and hit our backend (a
prerequisite to shadowing) the shadowed request should do so as well in
order to properly simulate the extra load.