Move *sync.WaitGroup.Add() outside of goroutine
There was actually a very subtle race condition with having the `Add`
within the gorouting.

If a process were to Start and then Stop very quickly it is possible
that a goroutine would have been started but not have ran `Add` wait, so
we don't wait for that goroutine to be `Done` correctly with `Wait`.

Pulling the call out to be before the goroutine starts fixes this
potential issue.